Updated view creation script to drop if exists first. Formatted date columns in evaluation lists.

+++ database/evaluations_view.sql
@@ -1,5 +1,7 @@
+use [CS6232-G1];
 GO
-DROP VIEW dbo.EvaluationScheduleView
+IF OBJECT_ID('dbo.EvaluationScheduleView', 'V') IS NOT NULL
+    DROP VIEW dbo.EvaluationScheduleView;
 GO
 CREATE VIEW dbo.EvaluationScheduleView
 AS
